libcontainer: Skip cgroup creation for rootless containers with cgroupfs
Creating cgroups with cgroupfs as non-root is not possible. Before this
change, we used to enforce systemd as a cgroup manager for all rootless
containers, but that prevents the usage of rootless youki on systems
wuthout systemd. To make such usage possible, raise a warning and skip
cgroup-related activities instead. That matches the behavior of crun
(containers/crun#97).

Fixes #3144

Signed-off-by: Michal Rostecki <vadorovsky@disroot.org>